refactor(schema): share Parquet FILE type
